feat/refactor (controllers/users):changed names of the user files, GET,POST,DELETE working
